[system] move functions dealing with generation of describes into a dedicated source file
 Conflicts:
	system/src/system_update.cpp
