+++ system/inc/system_update.h
@@ -6,9 +6,7 @@
 #include "static_assert.h"
 #include "appender.h"
 #include "system_defs.h"
-#include "ota_flash_hal.h"
-
-using particle::Appender;
+#include "system_info.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -50,14 +48,6 @@ PARTICLE_STATIC_ASSERT(system_file_transfer_size, sizeof(system_file_transfer_t)
 bool system_fileTransfer(system_file_transfer_t* transfer, void* reserved=NULL);
 
 void system_lineCodingBitRateHandler(uint32_t bitrate);
-
-bool system_module_info(appender_fn appender, void* append_data, void* reserved=NULL);
-bool system_metrics(appender_fn appender, void* append_data, uint32_t flags, uint32_t page, void* reserved=NULL);
-bool append_system_version_info(Appender* appender);
-
-typedef enum {
-    MODULE_INFO_JSON_INCLUDE_PLATFORM_ID = 0x0001
-} module_info_json_flags_t;
 
 /**
  *
@@ -159,23 +149,6 @@ void system_pending_shutdown(System_Reset_Reason reason);
 int system_set_flag(system_flag_t flag, uint8_t value, void* reserved);
 int system_get_flag(system_flag_t flag, uint8_t* value,void* reserved);
 int system_refresh_flag(system_flag_t flag);
-
-/**
- * Formats the diagnostic data using an appender function.
- *
- * @param id Array of data source IDs. This argument can be set to NULL to format all registered data sources.
- * @param count Number of data source IDs in the array.
- * @param flags Formatting flags.
- * @param append Appender function.
- * @param append_data Opaque data passed to the appender function.
- * @param reserved Reserved argument (should be set to NULL).
- */
-int system_format_diag_data(const uint16_t* id, size_t count, unsigned flags, appender_fn append, void* append_data,
-        void* reserved);
-
-
-int system_info_get(hal_system_info_t* info, uint32_t flags, void* reserved);
-int system_info_free(hal_system_info_t* info, uint32_t flags, void* reserved);
 
 #ifdef __cplusplus
 }
